+++ djpcms/forms/__init__.py
@@ -1,2 +1,5 @@
 from forms import *
+from fields import *
+from autocomplete import *
+from widgets import *
 from utils import *