+++ djpcms/utils/__init__.py
@@ -1,4 +1,5 @@
-from django.utils.encoding import smart_str, force_unicode, smart_unicode
+from django.utils.encoding import smart_str, force_unicode
+from django.utils.encoding import smart_unicode, StrAndUnicode
 from django.utils.safestring import mark_safe
 from django.utils.html import escape
 from djpcms.utils.numbers import significant_format, significant